requirements.txt
lxml==4.5.0
Pillow==6.1.0
PyMySQL==0.9.2
qiniu==7.2.8
requests==2.21.0
simplejson==3.15.0
